gui: Run emulation in a separate thread from the GUI
Two main 'loop' functions:

 * main_emulation_loop - Run as a separate thread.
   * Time is kept by SDL_GetTicks ().

 * main_gui_loop - Run in the main thread.
   * Time is kept by vsync.

Additional changes:

 * The .run () function pointer now takes in an uint32_t instead of a
   double. Any unused time is stored and added to the next call.

 * A mutex is added around state.video_out_data to prevent tearing.

 * state.running is unset during time-one-minute, to prevent two emulation
   threads running at the same time. This also mutes the sound while the
   minute is running.

+++ Source/snepulator.h
@@ -38,7 +38,7 @@ typedef struct Snepulator_State_s {
     char *cart_filename;
 
     /* Console API */
-    void (*run) (double ms);
+    void (*run) (uint32_t ms);
     void (*audio_callback) (void *userdata, uint8_t *stream, int len);
     uint32_t (*get_clock_rate) (void);
     bool (*get_int) (void);
